refactor:setAuthor,setFooter
